Added back in real tests. Removed hello world tests.
